Move questionnaire and response pages components to their parent directory
