+++ apps/smart-forms-react-app/src/components/Dashboard/QuestionnairePage/QuestionnairePageComponents/CreateNewResponseButton.tsx
@@ -1,11 +1,14 @@
 // @mui
 import { Button, CircularProgress } from '@mui/material';
 import React, { useContext, useState } from 'react';
-import { SelectedQuestionnaire } from '../../interfaces/Interfaces';
-import Iconify from '../Misc/Iconify';
-import { QuestionnaireProviderContext, QuestionnaireResponseProviderContext } from '../../App';
-import { LaunchContext } from '../../custom-contexts/LaunchContext';
-import { createQuestionnaireResponse } from '../../functions/QrItemFunctions';
+import { SelectedQuestionnaire } from '../../../../interfaces/Interfaces';
+import Iconify from '../../../Misc/Iconify';
+import {
+  QuestionnaireProviderContext,
+  QuestionnaireResponseProviderContext
+} from '../../../../App';
+import { LaunchContext } from '../../../../custom-contexts/LaunchContext';
+import { createQuestionnaireResponse } from '../../../../functions/QrItemFunctions';
 import { useNavigate } from 'react-router-dom';
 
 interface Props {
