+++ apps/smart-forms-react-app/src/components/Dashboard/QuestionnairePage/QuestionnairePageComponents/ViewExistingResponsesButton.tsx
@@ -1,14 +1,17 @@
 // @mui
 import { Button, CircularProgress } from '@mui/material';
 import React, { useContext, useMemo } from 'react';
-import Iconify from '../Misc/Iconify';
-import { LaunchContext } from '../../custom-contexts/LaunchContext';
-import { getClientBundlePromise, getResponsesFromBundle } from '../../functions/DashboardFunctions';
+import Iconify from '../../../Misc/Iconify';
+import { LaunchContext } from '../../../../custom-contexts/LaunchContext';
+import {
+  getClientBundlePromise,
+  getResponsesFromBundle
+} from '../../../../functions/DashboardFunctions';
 import { useQuery } from '@tanstack/react-query';
 import { Bundle, QuestionnaireResponse } from 'fhir/r5';
-import { SelectedQuestionnaireContext } from '../../custom-contexts/SelectedQuestionnaireContext';
+import { SelectedQuestionnaireContext } from '../../../../custom-contexts/SelectedQuestionnaireContext';
 import { useNavigate } from 'react-router-dom';
-import { SourceContext } from '../../Router';
+import { SourceContext } from '../../../../Router';
 
 function ViewExistingResponsesButton() {
   const { selectedQuestionnaire, setExistingResponses } = useContext(SelectedQuestionnaireContext);
