+++ apps/smart-forms-react-app/src/components/Dashboard/ResponsesPage/ResponsesPageComponents/BackToQuestionnairesButton.tsx
@@ -1,8 +1,8 @@
 // @mui
 import { Button } from '@mui/material';
 import React, { useContext } from 'react';
-import Iconify from '../Misc/Iconify';
-import { SelectedQuestionnaireContext } from '../../custom-contexts/SelectedQuestionnaireContext';
+import Iconify from '../../../Misc/Iconify';
+import { SelectedQuestionnaireContext } from '../../../../custom-contexts/SelectedQuestionnaireContext';
 import { useNavigate } from 'react-router-dom';
 
 function BackToQuestionnairesButton() {
