+++ apps/smart-forms-react-app/src/components/Dashboard/ResponsesPage/ResponsesPageComponents/OpenResponseButton.tsx
@@ -1,18 +1,21 @@
 // @mui
 import { Button, CircularProgress } from '@mui/material';
 import React, { useContext, useMemo, useState } from 'react';
-import Iconify from '../Misc/Iconify';
-import { LaunchContext } from '../../custom-contexts/LaunchContext';
-import { SelectedResponse } from '../../interfaces/Interfaces';
-import { QuestionnaireProviderContext, QuestionnaireResponseProviderContext } from '../../App';
+import Iconify from '../../../Misc/Iconify';
+import { LaunchContext } from '../../../../custom-contexts/LaunchContext';
+import { SelectedResponse } from '../../../../interfaces/Interfaces';
+import {
+  QuestionnaireProviderContext,
+  QuestionnaireResponseProviderContext
+} from '../../../../App';
 import { useQuery } from '@tanstack/react-query';
 import { Bundle, Questionnaire } from 'fhir/r5';
 import {
   getFormsServerAssembledBundlePromise,
   getFormsServerBundleOrQuestionnairePromise,
   getReferencedQuestionnaire
-} from '../../functions/DashboardFunctions';
-import { SourceContext } from '../../Router';
+} from '../../../../functions/DashboardFunctions';
+import { SourceContext } from '../../../../Router';
 import { useNavigate } from 'react-router-dom';
 
 interface Props {
