+++ apps/smart-forms-react-app/src/components/Dashboard/ResponsesPage/ResponsesPageComponents/ResponseListFeedback.tsx
@@ -8,7 +8,7 @@ import {
   TableRow,
   Typography
 } from '@mui/material';
-import { SourceContext } from '../../Router';
+import { SourceContext } from '../../../../Router';
 
 interface Props {
   isEmpty: boolean;
