+++ apps/smart-forms-react-app/src/components/Dashboard/ResponsesPage/ResponsesPageComponents/ResponseListToolbar.tsx
@@ -1,12 +1,12 @@
 import { IconButton, Tooltip, Typography } from '@mui/material';
-import Iconify from '../Misc/Iconify';
-import { StyledRoot } from '../Questionnaires/QuestionnaireListToolbar.styles';
-import { ResponseListItem } from '../../interfaces/Interfaces';
+import Iconify from '../../../Misc/Iconify';
+import { StyledRoot } from '../../QuestionnairePage/QuestionnairePageComponents/QuestionnaireListToolbar.styles';
+import { ResponseListItem } from '../../../../interfaces/Interfaces';
 import dayjs from 'dayjs';
 import { useContext } from 'react';
-import { SelectedQuestionnaireContext } from '../../custom-contexts/SelectedQuestionnaireContext';
-import { LaunchContext } from '../../custom-contexts/LaunchContext';
-import { constructName } from '../../functions/LaunchContextFunctions';
+import { SelectedQuestionnaireContext } from '../../../../custom-contexts/SelectedQuestionnaireContext';
+import { LaunchContext } from '../../../../custom-contexts/LaunchContext';
+import { constructName } from '../../../../functions/LaunchContextFunctions';
 
 interface Props {
   selected: ResponseListItem | undefined;
